[2.2][Feat] - Work on new cohort creation system
