+++ src/components/Cohort/CreationCohort/DiagramView/components/GroupCard/components/CriteriaCardContent/styles.ts
