+++ src/components/Cohort/CreationCohort/DiagramView/components/GroupCard/components/CriteriaRightPanel/CCAM/components/Hierarchy/CCAMHierarchy.tsx
