+++ src/components/Cohort/CreationCohort/DiagramView/components/GroupCard/components/CriteriaRightPanel/Cim10Form/components/Form/Cim10Form.tsx
