+++ src/components/Cohort/CreationCohort/DiagramView/components/GroupCard/components/CriteriaRightPanel/Cim10Form/styles.ts
