+++ src/components/Cohort/CreationCohort/DiagramView/components/GroupCard/components/CriteriaRightPanel/DemographicFrom/DemographicFrom.tsx
