+++ src/components/Cohort/CreationCohort/DiagramView/components/GroupCard/components/CriteriaRightPanel/GHM/components/Hierarchy/GhmHierarchy.tsx
